Add student project.
Well, it seems that to unify the process by using template programming
idea require lots of changes, so I add a 2 new packages. The promise
package includes all the base interfaces and structures that is used to
build other projects. The student package is the one that is used to
test this idea.